Package verified module info into lqhi file
The new Intr structure contains information about a verified module. It
is encoded and saved into a .lqhi file next to the source .hs or .lhs
file.

Currently this saving is done via a hacked-in testIntrGen function in
GhcInterface, which does the processing and prints the results as
debugging information. Eventually this will be tidied up and separated
out properly.