Set up test for POC of loading .lqhi data
LiquidHaskell now attempts to load and display test data for the .lqhi
file of the target module, if one exists.

Right now it fails to look up type constructors that were resolved to
ones defined in non-exported modules (eg. Data.Set.Set is
Data.Set.Base.Set; see tests/pos/ListConcat.hs). To fix this requires
modifying how type lookup happens when converting from the .lqhi
BareTypes to SpecTypes. Because it'd be awful to add in a *third* method
of converting from BareType -> SpecType in Bare, and because this code
is messy as it is, I'm going to attempt to overhaul this code and unify
it into a single, consistent interface. Wish me luck!

One change that's been made here is to the Symbolic instance for TyCon.
The previous implementation simply returned a qualifiedNameSymbol over
the TyCon's Name. Unfortunately, GHC's nameModule_maybe erroneously
returns "GHC.Types" for built-in type constructors like [], which causes
the .lqhi code to fail to resolve the symbol when it tries to look it
up. The Symbolic instance for TyCon has been modified to detect these
cases and return a non-prefixed Symbol, instead of returning an
incorrect result, for these built-in types.

Note that, if you want to test the .lqhi file loading code, you'll need
to run liquid twice on the test file: once to generate the .lqhi file,
and once to load it back in.