Naive modular verification
Similar to GhcMake. Say A depends on B, and B depends on C and D. C
will be verified first, then D, then B, and finally A. Each module
builds off of the information generated by verifying the previous
module.

Actually, for now, this is a naive implementation. The "information"
generated is just the BareSpec of the module, so each module has to go
back and reload all of its dependencies. The code is structured so
that it can be extended to support proper intermediate information
saving/loading.