Serialize LiquidHaskell data structures
This adds support for serialization to various LiquidHaskell data
structures. The BTyCon structure is also introduced. The "cereal"
library is used for this; instances are (for now) consolidated in
Serialize.hs.

As part of this, the way TyConInfo represents a type's size function
is changed from an actual Haskell function to a plain data type. This
is always either one of two structures, so the new representation
makes sense, and is necessary for serialization to happen.