Extract inline annotations
This only *identifies* the functions marked as inlines: it doesn't lift
them to logic or do any application, yet.

I want to see if I can reimplement the lift-to-logic step with
TypecheckedSource instead of CoreBinds. It should be possible, and it
would be less fragile than using CoreBinds (which are subject to
optimization etc).