Lift inlines to logic using CoreBinds
I've determined that replacing CoreBinds with TypecheckedSource here is
doable, but is outside this project's scope. I know which direction to
go in, but it will require more research and planning, and would fit
better in a more general CoreBinds->TypecheckedSource project. Maybe
something for future work.

Instead, I'm reusing the existing CoreToLogic infrastructure to lift
inlines to logic.