Embed LiquidHaskell as a TC plugin
This is a big commit, as it represents the result of several days of
work (and multiple iterations in other branches). Essentially,
LiquidHaskell goes from a separate executable to a compiler pass
embedded in GHC. This includes Cabal integration.

The test runner doesn't work right now, as it's still configured to run
the `liquid` executable.