Large bundle of GSoC work
This huge commit is the unfortunate result of accidentally `rm -r`ing my
.git folder. This was originally broken up into many individual commits,
but those have unfortunately been lost. Serves me right for not
bothering to push my changes! It shouldn't happen again, now that I've
disabled `rm` in my shell and switched to the `trash` command.

I'll try to summarize the highlights:

- `liquidhaskell-lang` has been merged in with the main repository.
  Since `liquidhaskell` works as a GHC plugin now, the `-lang` split
  would require depending on both packages instead of just this one.
- General polish work, tightening things up, smoothing down edge cases,
  and improving the LH user experience.
- Reorganize elements of the codebase, moving some related modules into
  well-defined subsystems (`Pipeline` for the core verification pipeline
  code, etc).
- Reducing our dependency list by eliminating some easily-replaced
  packages.
- Replacing the cmdargs-based command line parsing with a custom one
  better suited to a GHC plugin. Besides the improvement in user
  experience, this lets us drop the large cmdargs dependency, replacing
  it with ~150 lines of straightforward parsingg code.
- Cutting out config options and other corners of the code that don't
  make sense with the new setup.
- Tightening up the measures implementation and lift-to-logic code in
  general.
- Breaking out a reusable base from the parsing code that the config
  parser and the TH parser can share, reducing redundant code and making
  the parser core cleaner.
- Rewriting the type parser (again), ending up with a simpler version.
  I'm still not satisfied with this area of the code, though.
- Adding sort checking and other checks to the Spec subsystem.
- Work around GHC issue with generating .dyn_o for .hsig files. See GHC
  Trac #10660, for which I've submitted a patch. This patch has been
  accepted and will be included in 7.10.3, if it exists; otherwise,
  we'll see the fix in 7.12. Either way, LiquidHaskell works around the
  issue by generating the .dyn_o files itself.
- Various other enhancements.

There's a lot here, and without the history, it's difficult to pull out
everything that's changed, so I'm sure I'm missing things here. This
also doesn't include the various experimental changes that I made and
then backed out, which aren't really visible in the lumped-together
changeset. Sorry.

Lesson learned: *always* push changes *somewhere*, even if I'm not
confident in the stability of the codebase as I work on it. Don't be
afraid to put in-progress work out there.